% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMoments.R
\name{logMomentEstimator}
\alias{logMomentEstimator}
\title{Log-Moments Estimator for the Mittag-Leffler Distribution.}
\usage{
logMomentEstimator(x, a = 0.05)
}
\arguments{
\item{x}{A vector of non-negative data.}

\item{alpha}{Confidence level is 1 - alpha.}
}
\value{
A named vector with entries (nu     delta      nuLo      nuHi   deltaLo   
deltaHi)
}
\description{
Tail and scale parameter of the Mittag-Leffler distribution are estimated
by matching with the first two empirical log-moments.
}
\examples{
logMomentEstimator(rml(n = 100000, scale = 0.03, tail = 0.99))
}
