% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMLE.R
\name{mlml}
\alias{mlml}
\title{Maximum Likelihood Estimation of the Mittag-Leffler distribution}
\usage{
mlml(data, ...)
}
\arguments{
\item{data}{Vector of class "numeric"}

\item{...}{Additional parameters passed on to \code{\link{optim}}.}
}
\value{
A named vector of class "numeric"
}
\description{
Optimizes the bivariate likelihood function via 
\code{\link{optim}}. Uses \code{\link{logMomentEstimator}}
for initial parameter values.
}
