\name{test_mlfv}
\alias{test_mlfv}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function test_mlfv}
\description{
This function reproduces all the figures in the paper entitled Computation of the Mittag-Leffler function and its derivatives, Fract. Calc. Appl. Anal. 5(2002), 491-518 by  R.Gorenflo,  J.Loutchko, and Yu. Luchko}
\usage{
test_mlfv(fig=5) } 
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fig}{an integer between 5 and 15}
}
\details{
No previous data is used.
}
\value{
Returns a \code{\link[base]{character}} object }
\references{Please see Computation of the Mittag-Leffler function and its derivatives. 
Fract. Calc. Appl. Anal. 5(2002), 491-518 by  R.Gorenflo,  J.Loutchko, and Yu. Luchko}
\author{Eduardo M. A. M. Mendes}
\examples{
test_mlfv(5)
}
\keyword{methods}% at least one, from doc/KEYWORDS
