% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MittagLeffleR.R
\name{MittagLeffleR}
\alias{MittagLeffleR}
\title{The MittagLeffleR package.}
\description{
Contains the Mittag-Leffler function, distributions and a
Log-Moment estimator.
}
\section{Mittag-Leffler Distribution}{

The functions \code{\link{dml}}, \code{\link{pml}}, \code{\link{qml}} 
and \code{\link{rml}} for probability density, cumulative probability,
quantiles and random number generation.
}

\section{Mittag-Leffler function}{

is implemented as \code{\link{mlf}}.
}

\section{Log-moments estimator}{

see \code{\link{logMomentEstimator}}.
}

